refactor(nav): DigestPage 迁入 core 包，Route 改 sealed 拿到穷尽检查
DigestPage 是 Route 十九个实现里唯一住在 ui 包的（#80 早于 Route 接口
16 天诞生的历史遗留），sealed 要求直接实现同包，这一个例外挡住了整个
接口。平移进 core 独立文件（类、hnDiscussionUrl、三个 toDigestPage
转换器与测试原样搬迁，仅改包名），Route 加 sealed 后：

- entryProvider 的 when 获得编译期穷尽检查，删掉 else 运行时兜底
- 新增路由漏接 entryProvider 从运行时 crash 变为编译错误
- 清理 Route.kt 与 Hiring 两处解释该例外的过时 KDoc

头部零等待渲染、收藏同一性、存量 externalId 兼容均不受影响，
DigestPageTest 三个守护用例原样通过（allTests + iosArm64 编译验证）。

Closes #107
