+++ shared/src/commonMain/kotlin/whl/trending/ai/core/App.kt
@@ -15,7 +15,6 @@ import whl.trending.ai.ui.settings.AppearanceScreen
 import whl.trending.ai.ui.settings.ColorLabScreen
 import whl.trending.ai.ui.settings.DataSourcesScreen
 import whl.trending.ai.ui.settings.SettingsScreen
-import whl.trending.ai.ui.digest.DigestPage
 import whl.trending.ai.ui.hiring.HiringScreen
 import whl.trending.ai.ui.digest.DigestScreen
 import whl.trending.ai.ui.subscribe.SubscribeScreen
@@ -89,10 +88,7 @@ data class Chat(val context: ChatContext?) : Route { override val screen = Scree
 /**
  * HN 招聘月度专题。[month] 为 'YYYY-MM'，null = 最新一期（默认入口不带月份）。
  *
- * 与 [whl.trending.ai.ui.digest.DigestPage] 不同，本路由**不兼任数据载体**：
- * 它不从列表带任何内容进来（专题内容整月一次拉取），因此按惯例声明在这里，
- * 而不是跟着页面住在 ui 包——DigestPage 住在别处是被迫的例外（见 Route 的 KDoc），
- * 不是可照抄的范式。
+ * 与 [DigestPage] 不同，本路由不兼任数据载体：专题内容整月一次拉取，不从列表带内容。
  */
 data class Hiring(val month: String? = null) : Route { override val screen = Screen.HIRING }
 
@@ -365,10 +361,6 @@ fun App() {
                                     // 未注册（如 iOS）——入口本应隐藏，兜底直接返回
                                     LaunchedEffect(Unit) { backStack.safePop() }
                                 }
-                            }
-
-                            else -> {
-                                error("Unknown route: $key")
                             }
                             }
                         }
