+++ shared/src/commonTest/kotlin/whl/trending/ai/core/DigestPageTest.kt
@@ -1,4 +1,4 @@
-package whl.trending.ai.ui.digest
+package whl.trending.ai.core
 
 import whl.trending.ai.data.model.FavoriteItem
 import kotlin.test.Test
